pyproject.toml
[project]
name = "ezi-thesis"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.2.0",
    "networkx>=3.6.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dateutil>=2.9.0.post0",
    "scikit-fuzzy>=0.5.0",
    "scipy>=1.16.3",
]
